Split page-layout-box and move to core/dom
