+++ extensions/amp-auto-ads/0.1/ad-tracker.js
@@ -16,7 +16,7 @@
 
 import {Services} from '../../../src/services';
 import {endsWith} from '../../../src/core/types/string';
-import {measurePageLayoutBox} from '../../../src/utils/page-layout-box';
+import {measurePageLayoutBox} from './measure-page-layout-box';
 import {user} from '../../../src/log';
 
 /** @const */
