+++ extensions/amp-auto-ads/0.1/measure-page-layout-box.js
@@ -1,0 +1,27 @@
+/**
+ * Copyright 2021 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import {Services} from '.../../../src/services';
+import {getPageLayoutBoxBlocking} from '../../../src/core/dom/page-layout-box';
+import {toWin} from '../../../src/types';
+
+/**
+ * @param {!Element} element
+ * @return {!Promise<!../layout-rect.LayoutRectDef>}
+ */
+export function measurePageLayoutBox(element) {
+  const vsync = Services.vsyncFor(toWin(element.ownerDocument.defaultView));
+  return vsync.measurePromise(() => getPageLayoutBoxBlocking(element));
+}
